refactor: move elements files
