+++ src/pages/Home.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Header from '../components/Header';
 import Content from 'antd/es/layout/layout';
 import SearchBar from '../components/SerchBar';
-import style from './styled';
+import style from './styles';
 import Products from '../../assets/products';
 import ProductCard from '../components/ProductCard';
